requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
numpy>=1.21.0
supabase>=2.0.0
python-dotenv>=1.0.0

